[dagster-dbt] Move new cloud integration to separate folder
